+++ src/compiler/compiler_internal.h
@@ -48,7 +48,7 @@ typedef uint16_t FileId;
 #define MAX_ALIGNMENT ((ArrayIndex)(((uint64_t)2) << 28))
 #define MAX_GENERIC_DEPTH 32
 #define MAX_PRIORITY 0xFFFF
-#define MAX_TYPE_SIZE UINT32_MAX
+#define MAX_TYPE_SIZE (2U << 30)
 #define MAX_GLOBAL_DECL_STACK (65536)
 #define MAX_MODULE_NAME 31
 #define MAX_MODULE_PATH 63